[C#] Move WaitForSamples to Connector class
