+++ csharp/Connector/Reader.cs
@@ -99,23 +99,6 @@ namespace RTI.Connector
         }
 
         /// <summary>
-        /// Waits until a sample is received or the specified time passed.
-        /// </summary>
-        /// <param name="timeoutMillis">
-        /// Timeout in milliseconds.
-        /// Use -1 to wait indefinitely for samples.
-        /// </param>
-        public void WaitForSamples(int timeoutMillis)
-        {
-            if (Disposed)
-                throw new ObjectDisposedException(nameof(Reader));
-            if (timeoutMillis < -1)
-                throw new ArgumentOutOfRangeException(nameof(timeoutMillis));
-
-            reader.WaitForSamples(timeoutMillis);
-        }
-
-        /// <summary>
         /// Releases all resource used by the <see cref="Reader"/> object.
         /// </summary>
         /// <remarks>
